[TEST] Add unit test for _with_timeout helper
- Added tests/test_with_timeout_unit.py to comprehensively test the _with_timeout helper in server.py.
- Tests cover success, timeout exceeded, exception propagation, disabled timeout, and grace period cleanup.
- Removed tests/test_server_timeout.py which had collection errors and interfered with coverage reporting.
